Safari-611.1.10/LayoutTests/media/media-play-promise-reject-play-notallowed-expected.txt


RUN(internals.setMediaElementRestrictions(mediaElement, "RequireUserGestureForVideoRateChange"))
RUN(mediaElement.src = findMediaFile("video", "content/test"))
RUN(mediaElement.play().then(failTest).catch(promiseRejected))
Promise rejected. OK
EXPECTED (error.name == 'NotAllowedError') OK
END OF TEST
